pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptdeploy"
version = "0.1.0"
description = "Deploy prompts, agents, skills, and MCP servers to multiple tools."
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[project.scripts]
promptdeploy = "promptdeploy.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["promptdeploy"]

[tool.coverage.report]
fail_under = 98
show_missing = true
